Reorganise formatTraderType util
