+++ src/traders/format-trader-type.js
@@ -1,12 +1,14 @@
+const { TRADER_TYPE } = require('../constants');
+
 const formatTraderType = traderType => {
   if (traderType === undefined) {
     return undefined;
   }
 
   switch (traderType) {
-    case 0:
+    case TRADER_TYPE.MAKER:
       return 'maker';
-    case 1:
+    case TRADER_TYPE.TAKER:
       return 'taker';
     default:
       throw new Error(`Unrecognised trader type: ${traderType}`);
