+++ src/traders/format-trader-type.test.js
